Algoritmo HRRN y modifico un par de semaforos

+++ KernelT/src/kernelConfig.h
@@ -28,9 +28,9 @@ char* Puerto();
 //ALGORITMO
 char* Algoritmo();
 //ESTIMACION INICIAL
-int Estimacion();
+double Estimacion();
 //ALFA
-int Alfa();
+double Alfa();
 //MULTIPROGRAMACION
 int Multiprogramacion();
 //RECURSOS
